drawing primitives, fix timer for good

+++ source/modules/timer/wrap_timer.cpp
@@ -43,7 +43,7 @@ int Wrap_Timer::GetTime(lua_State * L)
 
 int Wrap_Timer::Sleep(lua_State * L)
 {
-    s64 seconds = luaL_checknumber(L, 1);
+    float seconds = luaL_checknumber(L, 1);
 
     instance()->Sleep(seconds);
